fix: refactor GatewayCheckoutSupportsListener to leaner listener with common trait and services
