+++ src/EventListener/RecomputingListenerTrait.php
@@ -12,14 +12,9 @@ trait RecomputingListenerTrait
     private function computeChangeSet(EntityManagerInterface $em, object $entity): void
     {
         $uow = $em->getUnitOfWork();
-
         $meta = $em->getClassMetadata($entity::class);
 
-        if ($entity->getId()) {
-            $uow->recomputeSingleEntityChangeSet($meta, $entity);
-        } else {
-            $em->persist($entity);
-            $uow->computeChangeSet($meta, $entity);
-        }
+        $em->persist($entity);
+        $uow->computeChangeSet($meta, $entity);
     }
 }
